requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-wtf==1.2.1
werkzeug==2.3.7
email-validator==2.1.0.post1
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.0.1
python-pptx==0.6.22
gunicorn==21.2.0